+++ zscaler/zid/services/groups/groups.go
@@ -10,7 +10,7 @@ import (
 	"encoding/json"
 
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler"
-	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zidentity/services/common"
+	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zid/services/common"
 )
 
 const (