+++ zscaler/zid/services/resource_servers/resource_servers.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler"
-	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zidentity/services/common"
+	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zid/services/common"
 )
 
 const (